requirements.txt
# IPFS Kit Python Core Requirements
# ==================================

# Core dependencies from pyproject.toml
requests>=2.28.0
httpx>=0.24.0  # For async HTTP requests and IPFS API client
psutil>=5.9.0
pyyaml>=6.0
base58>=2.1.1
multiaddr>=0.0.9  # For libp2p multiaddress support
importlib-metadata>=4.12.0; python_version < '3.10'
python-magic>=0.4.27  # For file type detection
anyio>=3.7.0  # For async operations with backend flexibility
trio>=0.22.0  # Optional backend for anyio
cryptography>=38.0.0  # Required for libp2p
toml

# MCP Server dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
websockets>=11.0.3
pydantic>=2.5.0

# Additional commonly used optional dependencies
transformers>=4.20.0
huggingface-hub>=0.10.0
fsspec>=2023.3.0
requests-unixsocket>=0.3.0
pyarrow>=11.0.0
pandas>=1.3.0
numpy>=1.20.0

# IPLD support
ipld-car>=0.0.1
ipld-dag-pb>=0.0.1
dag-cbor>=0.2.0
multiformats>=0.3.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.3.0
isort>=5.12.0
ipfs-api>=0.1.0
